feat: provide fetChannelList through the ChannelListContext
